Refactored `swipe` event
Much simpler velocity calculation.

Use the direction from the new `getAbsoluteMovement()` method instead of
recalculating.

Fixes a bug where the direction for a right swipe would sometimes be reported
as a left swipe. This was due to the direction calculation that has been
replaced in the `swipe` event's `onEnd` method.