Ability to define a "keep swiping" trigger point
Allow an optional `keepSwipingPercent` to be passed in to the constructor that
lets you set a percentage (in terms of how far out the left- or rightmost edge
of the list item has moved across the width of the list) that will trigger a
swipe regardless of velocity. This allows a user to swipe out past x percent,
think for a while, and if they still want it to swipe out let go and have it
swipe out if past the minimum percent. Makes the swiping gestures behave more
like they do in native Android applications.

The default value for swipe percent is 0 if it is missing, which returns false
for the check.

Use the direction returned from the `getAbsoluteMovement()` method instead of
re-calculating.

Fixed a bug with the assignment of options in the constructor.